+++ src/cue/animationcuehandler.h
@@ -43,7 +43,7 @@ namespace PixelMaestro {
 			uint8_t* set_plasma_options(uint8_t section_num, uint8_t layer_num, float size, float resolution);
 			uint8_t* set_radial_options(uint8_t section_num, uint8_t layer_num, uint8_t resolution);
 			uint8_t* set_sparkle_options(uint8_t section_num, uint8_t layer_num, uint8_t threshold);
-			uint8_t* set_wave_options(uint8_t section_num, uint8_t layer_num, uint8_t skew);
+			uint8_t* set_wave_options(uint8_t section_num, uint8_t layer_num, int8_t skew);
 
 			// General Animation calls
 			uint8_t* set_colors(uint8_t section_num, uint8_t layer_num, Colors::RGB* colors, uint8_t num_colors, bool delete_old_colors = true);